Moved yup validation to a utils folder, cleaned up signUp.js
